+++ Examples/CSharp/Wallet/UserAsset_Example.cs
@@ -1,0 +1,33 @@
+namespace Binance.Spot.WalletExamples
+{
+    using System;
+    using System.Net;
+    using System.Net.Http;
+    using System.Threading.Tasks;
+    using Binance.Common;
+    using Binance.Spot;
+    using Binance.Spot.Models;
+    using Microsoft.Extensions.Logging;
+
+    public class UserAsset_Example
+    {
+        public static async Task Main(string[] args)
+        {
+            using var loggerFactory = LoggerFactory.Create(builder =>
+            {
+                builder.AddConsole();
+            });
+            ILogger logger = loggerFactory.CreateLogger<UserAsset_Example>();
+
+            HttpMessageHandler loggingHandler = new BinanceLoggingHandler(logger: logger);
+            HttpClient httpClient = new HttpClient(handler: loggingHandler);
+
+            string apiKey = "api-key";
+            string apiSecret = "api-secret";
+
+            var wallet = new Wallet(httpClient, apiKey: apiKey, apiSecret: apiSecret);
+
+            var result = await wallet.UserAsset();
+        }
+    }
+}